update routes methods & naming convention
