+++ backend/Routes/api/media.routes.js
@@ -1,6 +1,10 @@
 import express from "express";
 import upload from "../../Middlewares/multer.middleware.js";
-import { uploadImage } from "../../Controllers/media.controller.js";
+
+import {
+    uploadImage
+} from "../../Controllers/media.controller.js";
+
 
 const mediaRoutes = express.Router();
