+++ backend/Routes/api/notification.routes.js
@@ -1,17 +1,26 @@
 import express from "express";
-import { addComment, getNotifications, deleteComment, getComments, getReplies, likeProject, likeStatus, newNotification, allNotificationsCount } from "../../Controllers/notification.controller.js";
-import { authenticateUser } from "../../Middlewares/auth.middleware.js";
+
+import {
+    authenticateUser
+} from "../../Middlewares/auth.middleware.js";
+
+import {
+    getNotifications,
+    likeProject,
+    likeStatus,
+    newNotification,
+    allNotificationsCount
+} from "../../Controllers/notification.controller.js";
+
 
 const notificationRoutes = express.Router();
 
+// User protected routes
 notificationRoutes.post("/like", authenticateUser, likeProject);
-notificationRoutes.post("/like-status", authenticateUser, likeStatus);
-notificationRoutes.post("/comment", authenticateUser, addComment);
-notificationRoutes.post("/get-comments", getComments);
-notificationRoutes.post("/get-replies", getReplies);
-notificationRoutes.post("/delete-comment", authenticateUser, deleteComment);
+notificationRoutes.get("/like-status", authenticateUser, likeStatus);
 notificationRoutes.get("/new", authenticateUser, newNotification);
-notificationRoutes.post("/get", authenticateUser, getNotifications);
-notificationRoutes.post("/all-count", authenticateUser, allNotificationsCount);
+notificationRoutes.get("/get", authenticateUser, getNotifications);
+notificationRoutes.get("/count", authenticateUser, allNotificationsCount);
+
 
 export default notificationRoutes;
