+++ backend/Routes/api/subscriber.routes.js
@@ -1,11 +1,23 @@
 import express from "express";
-import { getAllSubscribers, subscribeEmail, unsubscribeEmail } from "../../Controllers/subscriber.controller.js";
-import { authenticateUser } from "../../Middlewares/auth.middleware.js";
+
+import { 
+    authenticateUser 
+} from "../../Middlewares/auth.middleware.js";
+
+import { 
+    getAllSubscribers, 
+    subscribeEmail, 
+    unsubscribeEmail 
+} from "../../Controllers/subscriber.controller.js";
+
 
 const subscriberRoutes = express.Router();
 
 subscriberRoutes.post("/subscribe", subscribeEmail);
-subscriberRoutes.post("/unsubscribe", unsubscribeEmail);
+subscriberRoutes.patch("/unsubscribe", unsubscribeEmail);
+
+// Protected routes
 subscriberRoutes.get("/all", authenticateUser, getAllSubscribers);
+
 
 export default subscriberRoutes;
