+++ backend/Routes/api/user.routes.js
@@ -1,12 +1,25 @@
 import express from 'express';
-import { getProfile, searchUser, updateProfile, updateProfileImg } from '../../Controllers/user.controller.js';
-import { authenticateUser } from '../../Middlewares/auth.middleware.js';
+
+import {
+    authenticateUser
+} from '../../Middlewares/auth.middleware.js';
+
+import {
+    getProfile,
+    searchUser,
+    updateProfile,
+    updateProfileImg
+} from '../../Controllers/user.controller.js';
+
 
 const userRoutes = express.Router();
 
-userRoutes.post("/search", searchUser);
-userRoutes.post("/profile", getProfile);
-userRoutes.post("/update-profile-img", authenticateUser, updateProfileImg);
-userRoutes.post("/update-profile", authenticateUser, updateProfile);
+userRoutes.get("/search", searchUser);
+userRoutes.get("/profile", getProfile);
+
+// Protected routes
+userRoutes.patch("/update-profile-img", authenticateUser, updateProfileImg);
+userRoutes.patch("/update-profile", authenticateUser, updateProfile);
+
 
 export default userRoutes;
