+++ frontend/src/common/filter-pagination-data.jsx
@@ -15,7 +15,7 @@ export const filterPaginationData = async ({ create_new_arr = false, state, data
     if (state !== null && !create_new_arr) {
         obj = { ...state, results: [...state.results, ...data], page: page }
     } else {
-        await axios.post(import.meta.env.VITE_SERVER_DOMAIN + countRoute, data_to_send, headers)
+        await axios.get(import.meta.env.VITE_SERVER_DOMAIN + countRoute, data_to_send, headers)
             .then(({ data: { totalDocs } }) => {
                 obj = { results: data, page: 1, totalDocs }
             })
